+++ server/Option.go
@@ -24,7 +24,7 @@ type S2SRegistryOption struct { //服务端面向服务端的收发设置
 
 type S2SRegistrantOption struct { //服务端面向服务端的收发设置
 	ResponseSendOption protocol.ResponseSendOption
-	RegistryN          uint
+	RegistryN          uint64
 	CandidateList      registrant.RegistryCandidateList
 	RetryNController   registrant.RetryNController
 	RequestProto       protocol.RequestProtocol
